+++ spl/src/metadata.rs
@@ -1,7 +1,6 @@
 use anchor_lang::{
-    context::CpiContext,
-    prelude::{ErrorCode, UncheckedAccount},
-    Accounts, Result, ToAccountInfos,
+    context::CpiContext, error::ErrorCode, prelude::UncheckedAccount, Accounts, Result,
+    ToAccountInfos,
 };
 use mpl_token_metadata::{
     state::{CollectionDetails, DataV2, TokenMetadataAccount},
@@ -54,7 +53,6 @@ pub fn bubblegum_set_collection_size<'info>(
 
 pub fn burn_edition_nft<'info>(
     ctx: CpiContext<'_, '_, '_, 'info, BurnEditionNft<'info>>,
-    collection_metadata: Option<Pubkey>,
 ) -> Result<()> {
     let ix = mpl_token_metadata::instruction::burn_edition_nft(
         ID,
@@ -90,47 +88,6 @@ pub fn burn_nft<'info>(
         *ctx.accounts.edition.key,
         *ctx.accounts.spl_token.key,
         collection_metadata,
-    );
-    solana_program::program::invoke_signed(
-        &ix,
-        &ToAccountInfos::to_account_infos(&ctx),
-        ctx.signer_seeds,
-    )
-    .map_err(Into::into)
-}
-
-#[deprecated(note = "internal instructions deprecated by Metaplex")]
-pub fn create_metadata_accounts_v2<'info>(
-    ctx: CpiContext<'_, '_, '_, 'info, CreateMetadataAccountsV2<'info>>,
-    data: DataV2,
-    is_mutable: bool,
-    update_authority_is_signer: bool,
-) -> Result<()> {
-    let DataV2 {
-        name,
-        symbol,
-        uri,
-        creators,
-        seller_fee_basis_points,
-        collection,
-        uses,
-    } = data;
-    let ix = mpl_token_metadata::instruction::create_metadata_accounts_v2(
-        ID,
-        *ctx.accounts.metadata.key,
-        *ctx.accounts.mint.key,
-        *ctx.accounts.mint_authority.key,
-        *ctx.accounts.payer.key,
-        *ctx.accounts.update_authority.key,
-        name,
-        symbol,
-        uri,
-        creators,
-        seller_fee_basis_points,
-        update_authority_is_signer,
-        is_mutable,
-        collection,
-        uses,
     );
     solana_program::program::invoke_signed(
         &ix,
@@ -597,18 +554,6 @@ pub struct BurnNft<'info> {
     pub token: UncheckedAccount<'info>,
     pub edition: UncheckedAccount<'info>,
     pub spl_token: UncheckedAccount<'info>,
-}
-
-#[deprecated(note = "internal instructions deprecated by Metaplex")]
-#[derive(Accounts)]
-pub struct CreateMetadataAccountsV2<'info> {
-    pub metadata: UncheckedAccount<'info>,
-    pub mint: UncheckedAccount<'info>,
-    pub mint_authority: UncheckedAccount<'info>,
-    pub payer: UncheckedAccount<'info>,
-    pub update_authority: UncheckedAccount<'info>,
-    pub system_program: UncheckedAccount<'info>,
-    pub rent: UncheckedAccount<'info>,
 }
 
 #[derive(Accounts)]